UIDATIMP-1153: Add summary at top of individual import job's log
